requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
astroid==3.3.8
    # via pylint
autopep8==2.3.1
    # via -r requirements.in
beautifulsoup4==4.12.3
    # via bs4
bs4==0.0.2
    # via -r requirements.in
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
colorzero==2.0
    # via gpiozero
dill==0.3.9
    # via pylint
gpiozero==2.0.1
    # via -r requirements.in
idna==3.10
    # via requests
isort==5.13.2
    # via pylint
lgpio==0.2.2.0
    # via -r requirements.in
mccabe==0.7.0
    # via pylint
numpy==2.2.1
    # via -r requirements.in
pillow==11.0.0
    # via -r requirements.in
platformdirs==4.3.6
    # via pylint
pycodestyle==2.12.1
    # via autopep8
pylint==3.3.3
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
rpi-gpio==0.7.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
smbus==1.1.post2
    # via -r requirements.in
soupsieve==2.6
    # via beautifulsoup4
spidev==3.6
    # via -r requirements.in
tomlkit==0.13.2
    # via pylint
urllib3==2.3.0
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
